backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-generativeai==0.7.2
faiss-cpu==1.8.0
numpy==1.26.4
python-multipart==0.0.9
pydantic==2.7.1
packaging>=23.0
